Move seven Utilities/Extensions/Interfaces files into Core, split two more
Batch 2c. All nine files read in full per invariant 8.

Straight moves (PatternExecutor first, then ColumnSplitUtilities which
calls it): PatternExecutor, ColumnSplitUtilities, NumericUtilities,
LanguageHeuristics (internal - already covered by Core's existing
InternalsVisibleTo("Text-Grab")/("Tests")), Extensions/NumberExtensions,
Extensions/StringBuilderExtensions (uses StringMethods, already in Core),
Interfaces/ITtsEngine. No call sites needed fixing.

Two splits:

- ProtocolUtilities: IsProtocolUri/TryParseProtocolUri and the Scheme
  const are pure and keep the name in Core. TryGetSafeProtocolFilePath
  (needs AutomationProfile) and EnsureProtocolRegistration (needs
  Registry + FileUtilities) - both still app-side - moved into a new
  Text-Grab/Utilities/ProtocolHandlerUtilities.cs. Updated call sites:
  App.xaml.cs and the impure half of Tests/ProtocolUtilitiesTests.cs
  (the pure-half tests are untouched and still compile against Core
  through the existing InternalsVisibleTo; migrating them to Tests.Core
  is Wave 7a, not this batch).

- ThirdPartyNoticeUtilities: the Packages catalog and its constants are
  pure and keep the name in Core. GetBuiltWithFilePath/
  GetNoticesDirectoryPath/GetNoticeTarget/Open* all need
  FileUtilities.GetExePath() and moved into a new
  Text-Grab/Utilities/ThirdPartyNoticeLauncher.cs. Updated the one call
  site, LicensesWindow.xaml.cs. All three existing tests touch only
  Packages, so ThirdPartyNoticeUtilitiesTests.cs moved to Tests.Core in
  this same commit (namespace updated to Text_Grab.Tests.Core to match
  that project's convention).

Gate: Tests.Core 17/17 (14 baseline + 3 migrated), Tests.Core.Windows
3/3, Tests 1490/7 skipped/0 failed (1493 baseline - 3 migrated) -
matches baseline net of the test migration.
