+++ Tests.Core/ThirdPartyNoticeUtilitiesTests.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using Text_Grab.Utilities;
 
-namespace Tests;
+namespace Text_Grab.Tests.Core;
 
 public class ThirdPartyNoticeUtilitiesTests
 {
