+++ Text-Grab.Core/Extensions/NumberExtensions.cs
