+++ Text-Grab.Core/Extensions/StringBuilderExtensions.cs
