+++ Text-Grab.Core/Interfaces/ITtsEngine.cs
