+++ Text-Grab.Core/Utilities/ColumnSplitUtilities.cs
