+++ Text-Grab.Core/Utilities/LanguageHeuristics.cs
