+++ Text-Grab.Core/Utilities/NumericUtilities.cs
