+++ Text-Grab.Core/Utilities/PatternExecutor.cs
